.history/requirements_20241211031903.txt
annotated-types
anyio
APScheduler
asgiref
attrs
bcrypt
blinker
cattrs
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cycler
decorator
Django
dnspython
email_validator
fastapi
Flask
flatbuffers
fonttools
geopandas
git-filter-repo
greenlet
h11
idna
itsdangerous
Jinja2
kiwisolver
MarkupSafe
matplotlib
mysql-connector-python
numpy
openmeteo_requests
openmeteo_sdk
packaging
pandas
passlib
Pillow
pip
platformdirs
py
pycparser
pydantic
pydantic[email]
pydantic_core
PyKrige
PyMySQL
pyogrio
pyparsing
pyproj
python-dateutil
python-multipart
pytz
requests
requests_cache
retry
retry_requests
scipy
shapely
six
sniffio
SQLAlchemy
sqlparse
starlette
typing_extensions
tzdata
tzlocal
url-normalize
urllib3
uvicorn[standard]
Werkzeug
schedule